Transparent texture (#67)
* Decompile 802D1730

* func_802D17E4

* Geo18_802D2360

* Geo18_802D2470

* Geo18_802D2520

* Cleanup

+++ src/memory.c
@@ -30,7 +30,7 @@ int SegmentedToVirtual(u32 addr)
 
 u32 MakePtr(u32 segment, u32 addr)
 {
-    u32 offset = (addr & 0x1FFFFFFF) - gSegmentTable[segment];
+    u32 offset = ((u32)addr & 0x1FFFFFFF) - gSegmentTable[segment];
 
     return (segment << 24) + offset;
 }
